refactor: 역 등록 및 삭제 기능 service 분리
