+++ src/main/java/subway/ui/StationController.java
@@ -2,7 +2,8 @@ package subway.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.application.StationService;
+import subway.application.DeleteStationService;
+import subway.application.SaveSectionService;
 import subway.dto.StationRequest;
 
 import java.net.URI;
@@ -11,15 +12,17 @@ import java.net.URI;
 @RequestMapping("/stations")
 public class StationController {
 
-    private final StationService stationService;
+    private final SaveSectionService saveSectionService;
+    private final DeleteStationService deleteStationService;
 
-    public StationController(StationService stationService) {
-        this.stationService = stationService;
+    public StationController(SaveSectionService saveSectionService, DeleteStationService deleteStationService) {
+        this.saveSectionService = saveSectionService;
+        this.deleteStationService = deleteStationService;
     }
 
     @PostMapping
     public ResponseEntity<Void> createStation(@RequestBody StationRequest stationRequest) {
-        Long lineId = stationService.saveSection(stationRequest);
+        Long lineId = saveSectionService.saveSection(stationRequest);
         return ResponseEntity
                 .created(URI.create("/lines/" + lineId))
                 .build();
@@ -27,7 +30,7 @@ public class StationController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
-        Long lineId = stationService.deleteStationById(id);
+        Long lineId = deleteStationService.deleteStationById(id);
         return ResponseEntity
                 .noContent()
                 .location(URI.create("/lines/" + lineId))
